requirements.txt
docx==0.2.4
Flask==3.0.3
nltk==3.9.1
pandas==2.2.3
pdfplumber==0.11.4
psycopg2==2.9.9
python-dotenv==1.0.1
python_docx==1.1.2
spacy==3.7.6
